requirements.txt
Flask==2.3.3
APScheduler==3.10.4
pytest==8.2.1
python-dotenv==1.1.1